Debug the label, build std. 2 datasets for ML, apply rolling trick
build_labels.ipynb: Modyfy the codes because the previous one adopts the
label of dataset done rolling trick. It doesn't fit to the logic, so
the notebook is executed at the server to make sure the procedure of
label building is correct in the logic of this repository. Honestly,
it's quite painful to maintain the procedure's logic when arranging the
old scripts.

ML_element_01.ipynb: According to the logic, now we use the built label
to pair with two datasets (elemental intensities and spectra). The
paired datasets are exported. In this workflow we focus on the elemental
intensities first. In order to have better generalization of the future
model, the data are normalized sample-wides by Centred log-ratio
transformation. Subsequently, the rolling trick is applied. At the end,
the conjunction matrix is generated. Details are recorded in it.

README.md: update the step.

results and data are synchronized to Geololar server.