Train models using raw data and make comparison to rolling data
Since we know the rolling trick doesn't give enough improvement based on
the error analysis, we want to know comparing to the elemental data
without rolling, how good the rolling data gives machines to overcome
the fragmentation.

The python bash scripts are used to submit works, building models using
the raw data, to the cluster.

ML_element_10.ipynb: evaluate the results and discuss the comparison.

Outputs:
 raw_dev_conjunction_matrix_rf_20210519.png
 raw_dev_conjunction_matrix_svc_20210519.png
 raw_dev_conjunction_matrix_lr_20210519.png
 raw_dev_recall_%_svc_20210519.png
 raw_pca+rf_grid_20210519.png
 raw_rf_grid_20210519.png
 raw_pca+svc_grid_20201121.png
 raw_svc_grid_20201121.png
 raw_svc_grid_20210519.csv
 raw_rf_grid_20210519.csv
 raw_lr_grid_combined_20210519.png
 raw_lr_grid_20210519.csv
 raw_rf_grid_20210518.csv
 raw_lr_grid_20210518.csv
	new file:   ML_element_10.ipynb
	new file:   grid_raw_lr.py
	new file:   grid_raw_rf.py
	new file:   grid_raw_svc.py
	new file:   submit_raw.sh