Update section depth and a minor modification
We found a bug that many bottom sections having the composite depth as
their section depth so I fix it and update it. Details are in
update_database.ipynb. The database on the aslee@10.110.16.10 is
updated.

grid_rf.py is modified to use data in training-dev-test structure. This
is a preparation for the workflows related to raw data and image-like
data.
	modified:   grid_rf.py
	modified:   update_database.ipynb